refactor: move Dialect definitions to IR
