+++ crates/belalang_ir/include/belalang_ir/CMakeLists.txt
@@ -1,6 +1,1 @@
-set(LLVM_TARGET_DEFINITIONS Ops.td)
-mlir_tablegen(Ops.h.inc -gen-op-decls)
-mlir_tablegen(Ops.cpp.inc -gen-op-defs)
-mlir_tablegen(Dialect.h.inc -gen-dialect-decls)
-mlir_tablegen(Dialect.cpp.inc -gen-dialect-defs)
-add_public_tablegen_target(BelalangIROpsIncGen)
+add_subdirectory(IR)
