+++ crates/belalang_ir/include/belalang_ir/IR/Ops.td
@@ -1,6 +1,6 @@
 #ifndef BELALANG_IR_OPS
 #define BELALANG_IR_OPS
 
-include "belalang_ir/Dialect.td"
+include "belalang_ir/IR/Dialect.td"
 
 #endif // BELALANG_IR_OPS
