+++ crates/belalang_ir/lib/Dialect.cpp
@@ -1,3 +1,3 @@
-#include <belalang_ir/Dialect.h>
+#include <belalang_ir/IR/Dialect.h>
 
 using namespace mlir::belalang;
